pyproject.toml
[project]
name = "z-explorer"
version = "0.4.10"
description = "AI Image Generation Without the UI Tax - Z-Image-Turbo + Qwen3-4B"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "aiohttp>=3.11.18",
    "art>=6.5",
    "bitsandbytes>=0.48.2",
    "loguru>=0.7.3",
    "pillow>=10.4.0",
    "pydantic>=2.11.3",
    "python-dotenv>=1.1.0",
    "python-fasthtml>=0.12.14",
    "questionary>=2.1.0",
    "requests>=2.32.3",
    "rich>=13.9.4",
    "term-image>=0.7.2",
    "websockets>=15.0.1",
    "torch>=2.0.0",
    "transformers>=4.40.0",
    "accelerate>=0.30.0",
    "huggingface-hub>=0.23.0",
    "sentencepiece>=0.2.0",
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "sse-starlette>=2.0.0",
    "httpx>=0.27.0",
    "gguf>=0.17.1",
    "hf-transfer>=0.1.9",
    "outlines>=1.2.9",
    # Git dependencies - these require `uv sync` (not pip/uvx)
    "sdnq @ git+https://github.com/Disty0/sdnq",
    "diffusers @ git+https://github.com/huggingface/diffusers",
]


[tool.hatch.metadata]
allow-direct-references = true

[project.scripts]
z-explorer = "z_explorer.cli:main"


[build-system]
requires = ["uv_build>=0.9.1,<0.10.0"]
build-backend = "uv_build"

[tool.hatch.build.targets.wheel]
packages = ["src/z_explorer"]

[tool.hatch.build.targets.wheel.force-include]
"src/z_explorer/gui/dist" = "z_explorer/gui/dist"



[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.11.7",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
addopts = "--cov=z_explorer --cov-report=term-missing"

[tool.coverage.run]
source = ["src/z_explorer"]
omit = ["*/tests/*", "*/__pycache__/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
]
# fail_under = 80  # Enable after Phase 5 when full test suite is complete

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[tool.uv.sources]
torch = [
  { index = "pytorch-cu128", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
torchvision = [
  { index = "pytorch-cu128", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
